+++ docs/release-notes.md
@@ -26,6 +26,10 @@ UNRELEASED
 
 ##### New features
 
+* Python API: `MotionResult` (returned from `detect_motion()` and
+  `wait_for_motion()`) now has a member `region` which indicates where in the
+  video the motion was detected.
+
 ##### Minor fixes and packaging fixes
 
 * `MotionResult` now defines `__nonzero__()`. This means you can write